+++ src/components/load.py
@@ -90,7 +90,7 @@ class LoadComponent(ComponentBase):
             # Draw load factor frame (outline)
             painter.setPen(QPen(Qt.white, 1.5))
             painter.setBrush(Qt.NoBrush)
-            painter.drawRect(indicator_x, indicator_y, indicator_width, indicator_height)
+            painter.drawRect(int(indicator_x), int(indicator_y), int(indicator_width), int(indicator_height))
             
             # Determine fill color based on load factor - use gold with varying brightness
             # Darker gold for lower load, brighter gold for higher load
@@ -115,7 +115,7 @@ class LoadComponent(ComponentBase):
             # Calculate y-position for fill (starting from bottom of indicator)
             fill_y = indicator_y + (indicator_height - fill_height)
             
-            painter.drawRect(indicator_x, fill_y, indicator_width, fill_height)
+            painter.drawRect(int(indicator_x), int(fill_y), int(indicator_width), int(fill_height))
             
         else:
             # When graphics are disabled, still handle selection highlight but not shadow
@@ -173,7 +173,7 @@ class LoadComponent(ComponentBase):
                 scale_factor = 1.0 / view.transform().m11()  # Get inverse of horizontal scale
         
         # Set font with size adjusted for current zoom level
-        font = QFont('Arial', 14 * scale_factor)
+        font = QFont('Arial', int(14 * scale_factor))
         painter.setFont(font)
         
         # Draw the demand text centered below the image
