+++ src/components/pond.py
@@ -35,13 +35,13 @@ class PondComponent(ComponentBase):
             shadow_y + shadow_height / 2,
             shadow_width / 2
         )
-        gradient.setColorAt(0, QColor(0, 0, 0, 255 * self.shadow_opacity))
+        gradient.setColorAt(0, QColor(0, 0, 0, int(255 * self.shadow_opacity)))
         gradient.setColorAt(1, QColor(0, 0, 0, 0))
         
         # Draw the shadow
         painter.setBrush(QBrush(gradient))
         painter.setPen(Qt.NoPen)
-        painter.drawEllipse(shadow_x, shadow_y, shadow_width, shadow_height)
+        painter.drawEllipse(int(shadow_x), int(shadow_y), int(shadow_width), int(shadow_height))
         
         # Restore painter state after drawing shadow
         painter.restore()
